formatting cleanup

+++ build-rpi4.sh
@@ -129,8 +129,11 @@ waitfor () {
     # waitforit file is written in the function "endfunc"
     touch /flag/wait.${FUNCNAME[1]}_for_${1}
     printf "%${COLUMNS}s\n" "${FUNCNAME[1]} waits for: ${1}    "
-    while read waitforit; do if [ "$waitforit" = done.${1} ]; then break; \
-    fi; done \
+    while read waitforit; do 
+    if [ "$waitforit" = done.${1} ]; 
+        then break; \
+    fi; 
+    done \
    < <(inotifywait  -e create,open,access --format '%f' --quiet /flag --monitor)
     printf "%${COLUMNS}s\n" "${FUNCNAME[1]} noticed: ${1} [X]" && rm -f /flag/wait.${FUNCNAME[1]}_for_${1}
 }
